requirements.txt
django > 3.0
djangorestframework
django-cors-headers
django-heroku
gunicorn
django-crispy-forms
pusher
six
